+++ backend/canisters/local_user_index/api/src/updates/mod.rs
@@ -12,6 +12,7 @@ pub mod c2c_delete_group;
 pub mod c2c_group_index;
 pub mod c2c_local_group_index;
 pub mod c2c_mark_events_migrated_to_stable_memory;
+pub mod c2c_notifications_index;
 pub mod c2c_notify_low_balance;
 pub mod c2c_notify_user_events;
 pub mod c2c_notify_user_index_events;