+++ backend/canisters/local_user_index/impl/src/guards.rs
@@ -1,26 +1,34 @@
 use crate::read_state;
 
-pub fn caller_is_user_index_canister() -> Result<(), String> {
-    if read_state(|state| state.is_caller_user_index_canister()) {
+pub fn caller_is_user_index() -> Result<(), String> {
+    if read_state(|state| state.is_caller_user_index()) {
         Ok(())
     } else {
-        Err("Caller is not the user_index canister".to_string())
+        Err("Caller is not the user_index".to_string())
     }
 }
 
 pub fn caller_is_user_index_or_group_index() -> Result<(), String> {
-    if read_state(|state| state.is_caller_user_index_canister() || state.is_caller_group_index_canister()) {
+    if read_state(|state| state.is_caller_user_index() || state.is_caller_group_index()) {
         Ok(())
     } else {
         Err("Caller is not the UserIndex or GroupIndex".to_string())
     }
 }
 
-pub fn caller_is_group_index_canister() -> Result<(), String> {
-    if read_state(|state| state.is_caller_group_index_canister()) {
+pub fn caller_is_group_index() -> Result<(), String> {
+    if read_state(|state| state.is_caller_group_index()) {
         Ok(())
     } else {
-        Err("Caller is not the group_index canister".to_string())
+        Err("Caller is not the group_index".to_string())
+    }
+}
+
+pub fn caller_is_notifications_index() -> Result<(), String> {
+    if read_state(|state| state.is_caller_notifications_index()) {
+        Ok(())
+    } else {
+        Err("Caller is not the notifications_index".to_string())
     }
 }