+++ backend/canisters/local_user_index/impl/src/updates/c2c_upgrade_community_canister_wasm.rs
@@ -1,4 +1,4 @@
-use crate::guards::caller_is_group_index_canister;
+use crate::guards::caller_is_group_index;
 use crate::{Data, RuntimeState, mutate_state, read_state};
 use canister_api_macros::update;
 use canister_tracing_macros::trace;
@@ -9,7 +9,7 @@ use tracing::info;
 use types::{BuildVersion, CanisterId, CanisterWasm, ChunkedCanisterWasm, Hash, UpgradeChunkedCanisterWasmResponse::*};
 use utils::canister::{should_perform_upgrade, upload_wasm_in_chunks};
 
-#[update(guard = "caller_is_group_index_canister", msgpack = true)]
+#[update(guard = "caller_is_group_index", msgpack = true)]
 #[trace]
 async fn c2c_upgrade_community_canister_wasm(args: Args) -> Response {
     let PrepareResult { this_canister_id, wasm } = match read_state(|state| prepare(&args, state)) {