+++ backend/canisters/local_user_index/CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 
 - Store blocked user mappings in LocalUserIndexes ([#7960](https://github.com/open-chat-labs/open-chat/pull/7960))
+- Store notification subscriptions in LocalUserIndexes ([#7967](https://github.com/open-chat-labs/open-chat/pull/7967))
 
 ## [[2.0.1748](https://github.com/open-chat-labs/open-chat/releases/tag/v2.0.1748-local_user_index)] - 2025-05-13